requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-nmap==0.7.1
Werkzeug==3.1.3
sqlalchemy
vulners==1.5.18
urllib3==1.26.16
pymongo==4.0.1
reportlab==3.5.0
transformers==4.37.2